requirements.txt
# Core dependencies for Issues Injection
requests>=2.31.0
jsonschema>=4.19.0

# Optional: for enhanced functionality
python-dotenv>=1.0.0  # For loading .env files with credentials
pydantic>=2.0.0       # For advanced validation if needed
